requirements.txt
astroid==1.6.6
Django==2.0.2
ikp3db==1.4.1
isort==4.3.21
jedi==0.11.1
lazy-object-proxy==1.4.2
mccabe==0.6.1
parso==0.1.1
pbr==5.4.3
pylint==1.8.1
pylint-django==0.8.0
pylint-flask==0.5
pylint-plugin-utils==0.6
pytz==2019.3
six==1.12.0
stevedore==1.31.0
virtualenv==16.2.0
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wrapt==1.11.2
